Response is got from API, data is then stored into classes
tvList is an array holding objects.
Each object is a tv show
the object has some properties and then the array of its seasons

the seasons array holds objects again that holds some properties
and an array of its episodes

so the objects in tvList have access all the way down to episodes.

only one array to deal with

the tvList array is stored in the state and so the containers can access it

Next move is to integrate the starting container