Updated App.js - Resize Data as currently too big
The app is not loading as it is getting 429 responses from the api.
Too many requests are being made. 
I have made the data set smaller so that it flies under the 429 response radar. 
App can be restructured later to only load top level data and load data in a set time out way but for now just have a smaller data set